Crate wasm_runner

Application runner for Hermes WASM components. This runs each wasm file as a separate anonymous application.

Structs

Arguments

An example Wasm Runner

Constants

DEFAULT_ENV_TIMEOUT_MS 🔒

A standard value assigned to ENV_TIMEOUT_MS when it’s not specified.

ENV_TIMEOUT_MS 🔒

A parameter identifier specifying the max number of milliseconds to run the test for.

Functions

collect_apps 🔒

Collects .wasm files in the current directory or sub-directories of the current directory. Then creates one-module applications out of each of them.

collect_modules 🔒

Collects .wasm files in the current directory or sub-directories of the current directory. Return a String module name along with each compiled [Module].

create_one_module_app 🔒

Create one-module application with temp directory VFS.

create_temp_dir_child 🔒

Create a temp subdirectory.

get_timeout_ms 🔒

Get the timeout value from env.

init_ipfs 🔒

Initialize the IPFS node

init_logger 🔒

Init the logger

main 🔒

main_internal 🔒

“Regular” main function returning a Result that is mapped to an ExitCode in main()